habe long time zu int gemacht überall habe noch meinen code an euren angepasst und babysteps in compilerinteraction erstellt statt in tddtmenu babysteps-konstruktor überarbeitet es klappt, dass die zeitanzeige bei refactoring zurückgesetzt wird habe das eine römische zahlen auskommentiert, weil ich beim testen dauernd durcheinandergekommen bin

+++ src/main/java/de/hhu/propra16/coastal/tddt/CatalogParser.java
@@ -107,7 +107,7 @@ public class CatalogParser extends DefaultHandler {
                     } catch (ParseException e) {
                         e.printStackTrace(); // TODO throw
                     }
-                    mExercise.addBabystepTime(time.getTime()/1000);
+                    mExercise.addBabystepTime((int) time.getTime()/1000);
                 }
                 break;
             case TIMETRACKING:
